Dodavanje kontrolera za upravljanje korisnicima
Related Work Items: #18

+++ AutoSalon/AutoSalon/Models/Racun.cs
@@ -7,6 +7,6 @@ namespace AutoSalon.Models
         [Key]
         public int RacunID { get; set; }
         public DateTime DatumIzdavanja { get; set; }
-        public float Iznos { get; set; }
+        public decimal Iznos { get; set; }
     }
 }
